refactor: move store and source URI parsing to pkg/config (RFC 0022)
The URI grammar is a documented, user-facing contract — local:, s3:bucket/prefix,
b2:, sftp://user@host:port/path, and the gdrive/onedrive keyword forms — but
parsing it was private to cmd/cloudstic, so a caller outside this module had to
reimplement it to act on a store or source URI, with nothing keeping the two
readings in agreement.

It moves to pkg/config rather than to pkg/open, because parsing is resolution
rather than construction, and because the split is what decides the import
cost. Seventeen call sites in cmd/cloudstic parse a URI purely to validate one
— `profile new`, `store new`, the TUI forms, the config tables — and none of
them connect to anything. Putting the grammar in pkg/open would mean that
validating a profiles file links the AWS and Google clients. pkg/config stays
at 65 dependencies and no SDK.

storeURIParts and sourceURIParts become config.StoreURI and config.SourceURI,
dropping the Parts suffix the way the provider packages dropped theirs in §3,
and stay spellable in cmd/cloudstic as type aliases.

The function references had to change — Go has no alias for a func, so
parseStoreURI becomes config.ParseStoreURI at each of its ~20 call sites.
Mechanical, and the compiler covers it, but it is why this commit is not
call-site-neutral the way the config-type move was.

Two things documented on the way out, because they were implicit while the
types were private and become a public contract now:

- Which fields a parse populates depends on the scheme, so the zero fields on
  a StoreURI are meaningful rather than missing.
- SourceURI.Host carries the *drive name*, not a network host, for the gdrive
  and onedrive schemes. That overloading is worth revisiting, but renaming a
  field is a change to the shape rather than the location, and does not belong
  in a move.

Part of #378
